pyproject.toml
[project]
name = "nasa-ads-mcp"
version = "0.1.0"
description = "MCP server for NASA Astrophysics Data System - search papers, track citations, export BibTeX"
authors = [
    {name = "Paula Coelho", email = "pcoelho@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.1.2",
    "ads>=0.12.5",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/prtc/nasa-ads-mcp"
Repository = "https://github.com/prtc/nasa-ads-mcp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nasa_ads_mcp"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
